# backend/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/checkout/preview": {
      "post": {
        "operationId": "preview",
        "summary": "Preview checkout",
        "description": "Preview order details, pricing, and delivery options before confirmation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutPreviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout preview generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "subtotal": {
                          "type": "number",
                          "example": 599.98
                        },
                        "shipping": {
                          "type": "number",
                          "example": 25
                        },
                        "tax": {
                          "type": "number",
                          "example": 62.5
                        },
                        "total": {
                          "type": "number",
                          "example": 687.48
                        },
                        "currency": {
                          "type": "string",
                          "example": "USD"
                        },
                        "deliveryOptions": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid cart or currency"
          },
          "401": {
            "description": "Invalid or missing JWT token"
          }
        },
        "tags": [
          "checkout"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/checkout/confirm": {
      "post": {
        "operationId": "confirm",
        "summary": "Confirm checkout",
        "description": "Confirm and process the checkout with payment and delivery details",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutConfirmDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "orderId": {
                          "type": "string",
                          "example": "507f1f77bcf86cd799439011"
                        },
                        "orderNumber": {
                          "type": "string",
                          "example": "ORD-2024-001"
                        },
                        "status": {
                          "type": "string",
                          "example": "PENDING"
                        },
                        "totalAmount": {
                          "type": "number",
                          "example": 687.48
                        },
                        "currency": {
                          "type": "string",
                          "example": "USD"
                        },
                        "paymentUrl": {
                          "type": "string",
                          "example": "https://payment.example.com/checkout/..."
                        },
                        "estimatedDelivery": {
                          "type": "string",
                          "example": "2024-01-20T00:00:00Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid payment method or delivery address"
          },
          "401": {
            "description": "Invalid or missing JWT token"
          }
        },
        "tags": [
          "checkout"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/orders": {
      "get": {
        "operationId": "getMyOrders",
        "summary": "Get user orders",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "confirmed",
                "processing",
                "ready_to_ship",
                "shipped",
                "out_for_delivery",
                "delivered",
                "completed",
                "cancelled",
                "refunded",
                "returned",
                "payment_failed"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "operationId": "getOrderDetails",
        "summary": "Get order details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/orders/{id}/cancel": {
      "post": {
        "operationId": "cancelOrder",
        "summary": "Cancel order",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/orders/recent": {
      "get": {
        "operationId": "getRecentOrders",
        "summary": "Get recent orders",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/orders/{id}/track": {
      "get": {
        "operationId": "trackOrder",
        "summary": "Track order",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/orders/{id}/rate": {
      "post": {
        "operationId": "rateOrder",
        "summary": "Rate order",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/orders/stats/summary": {
      "get": {
        "operationId": "getMyStatistics",
        "summary": "Get user order statistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/orders/{id}": {
      "get": {
        "operationId": "getOrderDetails",
        "summary": "Get order details (Admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/orders/{id}/notes": {
      "post": {
        "operationId": "addNotes",
        "summary": "Add admin notes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/cart/preview": {
      "post": {
        "operationId": "preview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/cart/guest/preview": {
      "post": {
        "operationId": "preview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestPreviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cart-guest"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tagadod API",
    "description": "Complete e-commerce platform for solar products with advanced analytics and multi-language support",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.tagadod.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CheckoutPreviewDto": {
        "type": "object",
        "properties": {
          "currency": {
            "type": "string",
            "example": "YER",
            "enum": [
              "YER",
              "SAR",
              "USD"
            ]
          },
          "couponCode": {
            "type": "string"
          }
        },
        "required": [
          "currency"
        ]
      },
      "CheckoutConfirmDto": {
        "type": "object",
        "properties": {
          "deliveryAddressId": {
            "type": "string",
            "example": "65abc123def456789"
          },
          "currency": {
            "type": "string",
            "example": "YER"
          },
          "paymentMethod": {
            "type": "string",
            "example": "COD",
            "enum": [
              "COD",
              "ONLINE"
            ]
          },
          "paymentProvider": {
            "type": "string"
          },
          "shippingMethod": {
            "type": "string",
            "example": "standard"
          },
          "customerNotes": {
            "type": "string"
          },
          "couponCode": {
            "type": "string"
          }
        },
        "required": [
          "deliveryAddressId",
          "currency",
          "paymentMethod"
        ]
      },
...
      "CancelOrderDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "تغيير رأيي في المنتج"
          }
        },
        "required": [
          "reason"
        ]
      },
      "RateOrderDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "example": 5,
            "minimum": 1,
            "maximum": 5
          },
          "review": {
            "type": "string"
          }
        },
        "required": [
          "rating"
        ]
      },
...
      "PreviewDto": {
        "type": "object",
        "properties": {
          "currency": {
            "type": "string",
            "example": "YER",
            "enum": [
              "YER",
              "SAR",
              "USD"
            ]
          },
          "couponCode": {
            "type": "string"
          }
        },
        "required": [
          "currency"
        ]
      },
...
      "GuestPreviewDto": {
        "type": "object",
        "properties": {
          "currency": {
            "type": "string",
            "example": "YER",
            "enum": [
              "YER",
              "SAR",
              "USD"
            ]
          },
          "couponCode": {
            "type": "string"
          },
          "deviceId": {
            "type": "string",
            "example": "device_unique_id_123"
          }
        },
        "required": [
          "currency",
          "deviceId"
        ]
      },
...
    }
  }
}